log trace_pipe only when debug mode
